refactor: Category 클래스를 menu패키지로 이동
