+++ src/main/java/domain/menu/Category.java
@@ -1,0 +1,17 @@
+package domain.menu;
+
+public enum Category {
+	CHICKEN("치킨"),
+	BEVERAGE("음료");
+
+	private final String name;
+
+	Category(final String name) {
+		this.name = name;
+	}
+
+	@Override
+	public String toString() {
+		return "[" + name + "]";
+	}
+}
