+++ src/main/java/domain/menu/Menu.java
@@ -1,22 +1,20 @@
 package domain.menu;
 
-import domain.Category;
-
 public class Menu {
-    private final int number;
-    private final String name;
-    private final Category category;
-    private final int price;
+	private final int number;
+	private final String name;
+	private final Category category;
+	private final int price;
 
-    public Menu(final int number, final String name, final Category category, final int price) {
-        this.number = number;
-        this.name = name;
-        this.category = category;
-        this.price = price;
-    }
+	public Menu(final int number, final String name, final Category category, final int price) {
+		this.number = number;
+		this.name = name;
+		this.category = category;
+		this.price = price;
+	}
 
-    @Override
-    public String toString() {
-        return category + " " + number + " - " + name + " : " + price + "원";
-    }
+	@Override
+	public String toString() {
+		return category + " " + number + " - " + name + " : " + price + "원";
+	}
 }
